==> 12_a_Analyze_DE-GSEA_parental_vs_not.Rmd <==
```{r}
library(tidyverse)
library(fgsea)
`%notin%` <- negate(`%in%`)
source('functions/gseaFunctions.R')
source('functions/deVolcano.R')
fig_path <- "./plots/parent_de_analysis/"
```

## Load data
```{r load saved data}
outs <- load_de_gsea_for_analysis('binom_fusers_p05')

mb231.res <- outs[['mb231.res']]
h1806.res <- outs[['h1806.res']]
gsea <- outs[['gsea']]
```


## Sig genes
```{r }
l2fc_cutoff = 0.2
pval_cutoff = 0.05

top.mb231 <- mb231.res %>% filter(log2FC > l2fc_cutoff, p_val_adj < pval_cutoff)  %>% pull(gene)
top.h1806 <- h1806.res %>% filter(log2FC > l2fc_cutoff, p_val_adj < pval_cutoff)  %>% pull(gene)

bot.mb231 <- mb231.res %>% filter(log2FC < -l2fc_cutoff, p_val_adj < pval_cutoff)  %>% pull(gene)
bot.h1806 <- h1806.res %>% filter(log2FC < -l2fc_cutoff, p_val_adj < pval_cutoff)  %>% pull(gene)

toppers <- intersect(top.mb231, top.h1806)
bottoms <- intersect(bot.mb231, bot.h1806)


cat(paste0("p-adj < ",pval_cutoff," and abs(log2FC) > ",l2fc_cutoff,"\n"))
cat(paste0("\nMDA-MB-231:\n",
" ",nrow(mb231.res)," total genes tested\n",
"  ", length(top.mb231)+length(bot.mb231)," significant genes\n",
"   - ",length(top.mb231)," sig up\n",
"   - ",length(bot.mb231)," sig down\n"))

cat(paste0("\nHCC1806:\n",
" ",nrow(h1806.res)," total genes tested\n",
"  ", length(top.h1806)+length(bot.h1806)," significant genes\n",
"   - ",length(top.h1806)," sig up\n",
"   - ",length(bot.h1806)," sig down\n"))

cat(paste0("\nShared:\n",
"  ", length(toppers)+length(bottoms)," shared sig genes\n",
"   - ",length(toppers)," shared up\n",
"   - ",length(bottoms)," shared down\n"))
```

```{r}
l2fc_cutoff = 0.2
pval_cutoff = 0.05

pval_cutoff_mb = 1

top.mb231 <- mb231.res %>% filter(log2FC > l2fc_cutoff, p_val_adj <= pval_cutoff_mb)  %>% pull(gene)
top.h1806 <- h1806.res %>% filter(log2FC > l2fc_cutoff, p_val_adj <= pval_cutoff)  %>% pull(gene)

bot.mb231 <- mb231.res %>% filter(log2FC < -l2fc_cutoff, p_val_adj <= pval_cutoff_mb)  %>% pull(gene)
bot.h1806 <- h1806.res %>% filter(log2FC < -l2fc_cutoff, p_val_adj <= pval_cutoff)  %>% pull(gene)


toppers2 <- intersect(top.mb231, top.h1806)
bottoms2 <- intersect(bot.mb231, bot.h1806)

genes_to_highlight <- c(toppers2, bottoms2)

sig_genes <- rbind(mb231.res %>% mutate(cellline='MDA-MB-231'), 
                   h1806.res %>% mutate(cellline='HCC1806')) %>% 
  arrange(cellline) %>% 
  filter(gene %in% genes_to_highlight) %>% 
  select(cellline, gene, log2FC) %>% pivot_wider(names_from = cellline, values_from = log2FC) %>% 
  rowwise() %>%  
  mutate(HCC1806=round(HCC1806, 3), 
         `MDA-MB-231`=round(`MDA-MB-231`, 3), 
         mean_log2FC = mean(HCC1806, `MDA-MB-231`)) %>% 
  arrange(desc(mean_log2FC))


cat(paste0("p-adj < ",pval_cutoff_mb," and abs(log2FC) > ",l2fc_cutoff,"\n"))
cat(paste0("\nMDA-MB-231:\n",
" ",nrow(mb231.res)," total genes tested\n",
"  ", length(top.mb231)+length(bot.mb231)," significant genes\n",
"   - ",length(top.mb231)," sig up\n",
"   - ",length(bot.mb231)," sig down\n"))

cat(paste0("\np-adj < ",pval_cutoff," and abs(log2FC) > ",l2fc_cutoff,"\n"))
cat(paste0("\nHCC1806:\n",
" ",nrow(h1806.res)," total genes tested\n",
"  ", length(top.h1806)+length(bot.h1806)," significant genes\n",
"   - ",length(top.h1806)," sig up\n",
"   - ",length(bot.h1806)," sig down\n"))

cat(paste0("\nShared:\n",
"  ", length(toppers2)+length(bottoms2)," shared sig genes\n",
"   - ",length(toppers2)," shared up\n",
"   - ",length(bottoms2)," shared down\n"))


#write.table(sig_genes, file='outdata/de_results/de_sig_genes/sig_genes_p05_parent.tsv', quote=FALSE, sep='\t', col.names = TRUE)
```
## Volcano plots
```{r}
label_size = 2.5
label_pad = 0.25
n_genes_top_display = 12
title_font = 16
plot_font = 14

ph <- DE_volcano_top_and_pick_geneset(res = h1806.res,
                        shared_genes = c(bottoms, toppers),
                        highlight_shared_genes = genes_to_highlight,
                        title=paste0('HCC1806 non-random parent vs. not parent'), 
                        p_cut = 0.05, 
                        l2fc_cut = 0.1, 
                        n_genes_top = n_genes_top_display,
                        gene_label_size = label_size,
                        label.padding = label_pad,
                        axis_title_font_size = plot_font,
                        plot_title_font_size = title_font,
                        xmax = 2,
                        ymin = 0,
                        ymax = 10,
                        pick_seed = 123,
                        top_color = 'grey70',#'#F8766D', #'#00BFC4',
                        colrs=c('dodgerblue','goldenrod','grey20'),
                        colrs_main = c('grey80','grey90'),
                        highlight_colrs = c('blue','red'),
                        highlight_shape = 17)


pm <- DE_volcano_top_and_pick_geneset(res = mb231.res, 
                        shared_genes = c(bottoms, toppers),
                        highlight_shared_genes = genes_to_highlight,
                        title=paste0('MDA-MB-231 non-random parent vs. not parent'), 
                        p_cut = 0.05, 
                        l2fc_cut = 0.1, 
                        n_genes_top = n_genes_top_display,
                        gene_label_size = label_size,
                        label.padding = label_pad,
                        axis_title_font_size = plot_font,
                        plot_title_font_size = title_font,
                        xmax = 2,
                        ymin = 0,
                        ymax = 2,
                        pick_seed = 123,
                        top_color = 'grey70',#'#F8766D', #'#00BFC4',
                        colrs=c('dodgerblue','goldenrod','grey20'),
                        colrs_main = c('grey80','grey90'),
                        highlight_colrs = c('blue','red'),
                        highlight_shape = 17) 



#SaveFigure(ph, 'h1806_volcano_bulk_parent_binom', type = "png", width=8, height=4, res=600)
#SaveFigure(pm, 'mb231_volcano_bulk_parent_binom', type = "png", width=8, height=4, res=600)


ph
pm

```

## GSEA plots
```{r}
library(scales)
source('functions/gseaFunctions.R')

n_plot= 20
p_thresh = 1
padj_thresh = 0.25
NES_min_thresh = 0
str.max = 50

hall <- dot_plot_gsea(gsea, 
              pick_gs='HALLMARK', 
              pval_threshold  = p_thresh, 
              padj_threshold  = padj_thresh, 
              NES_min = NES_min_thresh, 
              n_pathway_plot = n_plot, 
              min_celllines_in = 2, 
              gs_name_length = 1,
              data_title = 'non-random parent vs. not parent', 
              only_pos = FALSE
              ) + scale_y_discrete(labels = label_wrap(width = str.max))

gobp <- dot_plot_gsea(gsea, 
              pick_gs='GOBP', 
              pval_threshold  = p_thresh, 
              padj_threshold  = padj_thresh, 
              NES_min = NES_min_thresh, 
              n_pathway_plot = n_plot, 
              min_celllines_in = 2, 
              gs_name_length = 1,
              data_title = 'non-random parent vs. not parent', 
              only_pos = FALSE
              ) + scale_y_discrete(labels = label_wrap(width = str.max))


hall
gobp
#SaveFigure(gobp+theme(aspect.ratio=3), 'gsea_dotplot_bulk_parent_binom_gobp', type = "png", width=10, height=8, res=600)
#SaveFigure(hall, 'gsea_dotplot_bulk_parent_binom_hallmark', type = "png", width=7, height=4, res=600)
```

```{r}
gomf <- dot_plot_gsea(gsea, 
              pick_gs='GOMF', 
              pval_threshold  = p_thresh, 
              padj_threshold  = padj_thresh, 
              NES_min = NES_min_thresh, 
              n_pathway_plot = n_plot, 
              min_celllines_in = 2, 
              gs_name_length = 1,
              data_title = 'non-random parent vs. not parent', 
              only_pos = FALSE
              ) + scale_y_discrete(labels = label_wrap(width = str.max))

kegg <- dot_plot_gsea(gsea, 
              pick_gs='KEGG', 
              pval_threshold  = p_thresh, 
              padj_threshold  = padj_thresh, 
              NES_min = NES_min_thresh, 
              n_pathway_plot = n_plot, 
              min_celllines_in = 2, 
              gs_name_length = 1,
              data_title = 'non-random parent vs. not parent', 
              only_pos = FALSE
              ) + scale_y_discrete(labels = label_wrap(width = str.max))


gomf
kegg
```
